Create Contacts Book
Contacts in the book have name and [phone] number.

They can be ordered in ascending or descending order
by name or by number.
They can be searched by full name or number,
by the part of a name or a number or
by the start or the end of a name or a number.

Signed-off-by: Polina Shlepakova <polinashlepakova@gmail.com>